+++ src/gamspy/_symbols/equation.py
@@ -256,15 +256,15 @@ class Equation(gt.Equation, Symbol):
             if description != "":
                 self.description = description
 
-            previous_state = self.container.miro_protect
-            self.container.miro_protect = False
+            previous_state = self.container._options.miro_protect
+            self.container._options.miro_protect = False
             self.records = None
             self.modified = True
 
             # only set records if records are provided
             if records is not None:
                 self.setRecords(records, uels_on_axes=uels_on_axes)
-            self.container.miro_protect = previous_state
+            self.container._options.miro_protect = previous_state
 
         else:
             type = cast_type(type)
@@ -277,8 +277,8 @@ class Equation(gt.Equation, Symbol):
             else:
                 name = "e" + str(uuid.uuid4()).replace("-", "_")
 
-            previous_state = container.miro_protect
-            container.miro_protect = False
+            previous_state = container._options.miro_protect
+            container._options.miro_protect = False
             super().__init__(
                 container,
                 name,
@@ -319,7 +319,7 @@ class Equation(gt.Equation, Symbol):
             else:
                 self.container._synch_with_gams()
 
-            container.miro_protect = previous_state
+            container._options.miro_protect = previous_state
 
     def __hash__(self):
         return id(self)