+++ src/gamspy/_symbols/parameter.py
@@ -199,15 +199,15 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
             if description != "":
                 self.description = description
 
-            previous_state = self.container.miro_protect
-            self.container.miro_protect = False
+            previous_state = self.container._options.miro_protect
+            self.container._options.miro_protect = False
             self.records = None
             self.modified = True
 
             # only set records if records are provided
             if records is not None:
                 self.setRecords(records, uels_on_axes=uels_on_axes)
-            self.container.miro_protect = previous_state
+            self.container._options.miro_protect = previous_state
         else:
             if name is not None:
                 name = validation.validate_name(name)
@@ -217,8 +217,8 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
             else:
                 name = "p" + str(uuid.uuid4()).replace("-", "_")
 
-            previous_state = container.miro_protect
-            container.miro_protect = False
+            previous_state = container._options.miro_protect
+            container._options.miro_protect = False
             super().__init__(
                 container,
                 name,
@@ -244,7 +244,7 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
             else:
                 self.container._synch_with_gams()
 
-            container.miro_protect = previous_state
+            container._options.miro_protect = previous_state
 
     def __getitem__(
         self, indices: Sequence | str
@@ -261,7 +261,7 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
         # self[domain] = rhs
         domain = validation.validate_domain(self, indices)
 
-        if self._is_miro_input and self.container.miro_protect:
+        if self._is_miro_input and self.container._options.miro_protect:
             raise ValidationError(
                 f"Cannot assign to protected miro input symbol {self.name}. `miro_protect`"
                 " attribute of the container can be set to False to allow"
@@ -372,7 +372,7 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
         if (
             hasattr(self, "_is_miro_input")
             and self._is_miro_input
-            and self.container.miro_protect
+            and self.container._options.miro_protect
         ):
             raise ValidationError(
                 "Cannot assign to protected miro input symbols. `miro_protect`"